requirements.txt
opentelemetry-exporter-prometheus-remote-write
awslambdaric